chore(typescript): improved types to match implementations
