+++ packages/victory-core/src/types.d.ts
@@ -1,12 +1,3 @@
-import * as React from "react";
-import { AnimationEasing, AnimationStyle } from "./animationStyle";
-import {
-  NumberOrCallback,
-  PaddingProps,
-  VictoryThemeDefinition
-} from "./victory-theme/victory-theme";
-import { OriginType } from "./victory-label/victory-label";
-
 /**
  * This is the first parameter of a callback when a callback is used to
  * resolve the value of a property instead of a concrete value.
@@ -45,47 +36,7 @@ export type SliceNumberOrCallback<T, P = null> =
   | number
   | ((props: Omit<T, P>) => number);
 
-
 export type CoordinatesPropType = {
   x: number;
   y: number;
 };
-
-// #endregion
-
-// #region Victory Animation
-// #endregion
-
-// #region Victory Axis props
-
-// #endregion
-
-// #region Victory Label
-
-// #endregion
-
-// #region Victory Container
-
-// #endregion
-
-// #region Victory Clip Container
-
-// #endregion
-
-// #region Victory Accessibile Group
-
-// #endregion
-
-// #endregion
-
-// #region Victory Util
-
-// #endregion
-
-// #region Victory Portal
-
-// #endregion
-
-// #region Victory Primitives
-
-// #endregion
